apps/api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
langgraph
langgraph-checkpoint-postgres
pydantic-settings==2.2.1
psycopg2-binary
asyncpg
httpx==0.27.0
python-dotenv==1.0.1
PyGithub==2.3.0
sqlalchemy==2.0.29
pgvector==0.2.5
langchain-core==0.2.6
langchain-openai==0.1.7
redis==5.0.3
alembic==1.13.1
celery[redis]==5.4.0
flower==2.0.1
slowapi==0.1.9
structlog==24.1.0
python-jose[cryptography]==3.3.0
cryptography==42.0.5
asgi-correlation-id==4.1.0
pydantic==2.6.4
email-validator==2.1.1
fakeredis==2.21.1
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==4.1.0
aiosqlite==0.20.0
pyyaml==6.0.1
openai>=1.0.0APScheduler==3.10.4
